deployment/gradio_app/requirements.txt
gradio>=3.44.0
torch>=2.0.0
torchvision>=0.15.0
pillow>=9.5.0
